+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.3
+
+* Some helpful extensions
+
+## 1.0.2
+
+* Actually set the baseurl
+
 ## 1.0.0
 
 * Initial Release